backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.30.6"
fastapi = "^0.112.0"
asyncpg = "^0.29.0"
arq = "^0.26.0"
config = "^0.5.1"
aerich = "0.7.2"
tortoise-orm = "^0.21.3"
setuptools = "^72.2.0"
python-dotenv = "^1.0.1"
aiohttp = ">=3.9.5,<3.10"


[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
